PERSISTENCE ACTUALLY WORKS! NO ERRORS
the serializing seemed to be looping, so i just made it so that the only thing that's saved is an ID, which is then used to grab the channel/message necessary! this also reduces storage by a significant amount.

i added in the structure for count/chain/convo, and nearly fully implemented counting, with all(?) the features of Caretaker Archive

ive been using StringBuilder a lot more, and ive replaced a few things that used lists before. i haven't tested them for performance, but they looped a LOT so i would imagine that it's some good performance gain :)

things that rely on the client are less static based, and on init they're all put in Ready(), which is a method that runs when the client is ready. who knew it was that easy 🙂