ReactionSubscribe class + NEW HELP COMMAND!!
oh okay so this is the best thing ive done in my life
ReactionSubscribe implements IDisposable, which should make it much more resistant to random memory leaks n stuff. it also just makes it easier to code with/it produces much nicer looking, clean code. So freakin epic
reworked the help command to return "(string, string)[]?" which is a nullable array of string tuples. that's weird. but it works so well. the first string is the genre of all the commands, then the second string is all of the commands parsed into one string for display
gotta make it look better tho

ALSO i need to separate this CaretakerCore stuff. for now it's fine, the main thing i wanted to do was to separate the code stuff from the main project (get rid of any MainHook calls, put those methods in MainHook itself or just reconfigure them)